@@ -5,12 +5,9 @@
   "dependencies": {
     "@shopify/polaris": "^1.5.1",
     "axios": "^0.16.2",
-    "babel-eslint": "^8.0.1",
-    "cypress": "^1.0.3",
-    "eslint-plugin-react": "^7.4.0",
-    "firebase": "^4.1.3",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "firebase": "^7.14.3",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-dropzone": "^3.13.3",
     "react-ga": "^2.3.5",
     "react-router-dom": "^4.1.1",
@@ -21,21 +18,34 @@
     "stripe": "^5.1.1"
   },
   "devDependencies": {
-    "enzyme": "^2.9.1",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-node": "^5.2.1",
-    "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "jest": "^20.0.4",
-    "react-scripts": "1.0.10",
-    "react-test-renderer": "^15.6.1"
+    "cypress": "^4.5.0",
+    "eslint-config-standard": "^14.1.1",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.19.0",
+    "eslint-plugin-standard": "^4.0.1",
+    "react-scripts": "^3.4.1",
+    "start-server-and-test": "^1.11.0"
   },
   "scripts": {
     "start": "node node_modules/react-scripts/scripts/start.js",
     "build": "node node_modules/react-scripts/scripts/build.js",
     "eject": "node node_modules/react-scripts/scripts/eject.js",
     "coverage": "yarn run jest -- --coverage",
-    "tests": "cypress open"
+    "cypress:open": "npx cypress open",
+    "cypress": "start-server-and-test start http://localhost:3000 cypress:open"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
