@@ -1,4 +1,9 @@
-import firebase from 'firebase'
+import * as firebase from 'firebase'
+// import firebase from 'firebase/app';
+// import 'firebase/auth';
+// import 'firebase/firestore';
+// import 'firebase/performance';
+// import 'firebase/storage';
 
 const config =
   process.env.NODE_ENV === 'production'
